Fixed "Tablet" Spacing and Mobile Nav Menu Issue
The "tablet" now has a scrollable "main" which allows for the content to be fully displayed while also added to the "tablet" effect. The Header and Footer are both fixed within the "tablet". The Mobile Navigation Menu is now displaying correctly.

Now working on getting a second page added and to have a pop up contact form. If the contact form isn't achieved easily, I will made a contact page and add a form to it, then work on a pop up form another time.